update: moved sqlutil to util/sql
