+++ util/sql/sqlutil_test.go
@@ -1,4 +1,4 @@
-package sqlutil
+package sql
 
 import (
 	"github.com/stretchr/testify/assert"
